+++ Mocolamma/Views/ModelInspectorView.swift
@@ -120,13 +120,7 @@ struct ModelInspectorView: View {
     }
     
     var body: some View {
-        let copyIconName = {
-            if #available(macOS 15.0, iOS 18.0, *) {
-                return "document.on.document"
-            } else {
-                return "doc.on.doc"
-            }
-        }()
+        let copyIconName = SFSymbol.copy
         
         ScrollView {
             VStack(alignment: .leading, spacing: 10) {